fix(lazy): do not auto-install plugins
When switching machines this always regenerates the lazy-lock.json. This
causes issues when another machine added a plugin and also updated the
pinned plugin versions. In that case, the pulling machine will install a
missing plugin, regenerate the lazy-lock.json with the currently
INSTALLED plugins and thus overriding the also included change of
updating other plugins.